privatize credentials
