+++ sdk/identity/azure-identity/azure/identity/_credentials/client_credential.py
@@ -1,0 +1,77 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# ------------------------------------
+from .._authn_client import AuthnClient
+from .._base import ClientSecretCredentialBase, CertificateCredentialBase
+from .._constants import Endpoints
+
+try:
+    from typing import TYPE_CHECKING
+except ImportError:
+    TYPE_CHECKING = False
+
+if TYPE_CHECKING:
+    # pylint:disable=unused-import,ungrouped-imports
+    from typing import Any, Mapping
+    from azure.core.credentials import AccessToken
+
+
+class ClientSecretCredential(ClientSecretCredentialBase):
+    """
+    Authenticates as a service principal using a client ID and client secret.
+
+    :param str client_id: the service principal's client ID
+    :param str secret: one of the service principal's client secrets
+    :param str tenant_id: ID of the service principal's tenant. Also called its 'directory' ID.
+    """
+
+    def __init__(self, client_id, secret, tenant_id, **kwargs):
+        # type: (str, str, str, Mapping[str, Any]) -> None
+        super(ClientSecretCredential, self).__init__(client_id, secret, tenant_id, **kwargs)
+        self._client = AuthnClient(Endpoints.AAD_OAUTH2_V2_FORMAT.format(tenant_id), **kwargs)
+
+    def get_token(self, *scopes, **kwargs):  # pylint:disable=unused-argument
+        # type: (*str, **Any) -> AccessToken
+        """
+        Request an access token for `scopes`.
+
+        :param str scopes: desired scopes for the token
+        :rtype: :class:`azure.core.credentials.AccessToken`
+        :raises: :class:`azure.core.exceptions.ClientAuthenticationError`
+        """
+        token = self._client.get_cached_token(scopes)
+        if not token:
+            data = dict(self._form_data, scope=" ".join(scopes))
+            token = self._client.request_token(scopes, form_data=data)
+        return token
+
+
+class CertificateCredential(CertificateCredentialBase):
+    """
+    Authenticates as a service principal using a certificate.
+
+    :param str client_id: the service principal's client ID
+    :param str tenant_id: ID of the service principal's tenant. Also called its 'directory' ID.
+    :param str certificate_path: path to a PEM-encoded certificate file including the private key
+    """
+
+    def __init__(self, client_id, tenant_id, certificate_path, **kwargs):
+        # type: (str, str, str, Mapping[str, Any]) -> None
+        self._client = AuthnClient(Endpoints.AAD_OAUTH2_V2_FORMAT.format(tenant_id), **kwargs)
+        super(CertificateCredential, self).__init__(client_id, tenant_id, certificate_path, **kwargs)
+
+    def get_token(self, *scopes, **kwargs):  # pylint:disable=unused-argument
+        # type: (*str, **Any) -> AccessToken
+        """
+        Request an access token for `scopes`.
+
+        :param str scopes: desired scopes for the token
+        :rtype: :class:`azure.core.credentials.AccessToken`
+        :raises: :class:`azure.core.exceptions.ClientAuthenticationError`
+        """
+        token = self._client.get_cached_token(scopes)
+        if not token:
+            data = self._get_request_data(*scopes)
+            token = self._client.request_token(scopes, form_data=data)
+        return token
