+++ sdk/identity/azure-identity/azure/identity/aio/__init__.py
@@ -5,7 +5,7 @@
 import os
 
 from .._constants import EnvironmentVariables
-from .credentials import (
+from ._credentials import (
     CertificateCredential,
     ChainedTokenCredential,
     ClientSecretCredential,
